third_party/blink/web_tests/crypto/subtle/aes-gcm/cloneKey-expected.txt

Tests structured cloning of AES-GCM keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: 5c4b010910031030112233445566778899aabbccddeeffa0000000014b
PASS: Cloned key exported data should be [30112233445566778899aabbccddeeff] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: 5c4b010920032000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0fa0000000014b
PASS: Cloned key exported data should be [00112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: 5c4b010910451030112233445566778899aabbccddeeffa0000000014b
PASS: Cloned key exported data should be [30112233445566778899aabbccddeeff] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: 5c4b010920452000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0fa0000000014b
PASS: Cloned key exported data should be [00112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: 5c4b010910c3011030112233445566778899aabbccddeeffa0000000014b
PASS: Cloned key exported data should be [30112233445566778899aabbccddeeff] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is true
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is true
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: 5c4b010920c3012000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0fa0000000014b
PASS: Cloned key exported data should be [00112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f] and was


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: 5c4b010910021030112233445566778899aabbccddeeffa0000000014b


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "encrypt"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "encrypt"
Serialized key bytes: 5c4b010920022000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0fa0000000014b


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: 5c4b010910441030112233445566778899aabbccddeeffa0000000014b


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "decrypt,wrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "decrypt,wrapKey"
Serialized key bytes: 5c4b010920442000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0fa0000000014b


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 128
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 128
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: 5c4b010910c2011030112233445566778899aabbccddeeffa0000000014b


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "AES-GCM"
PASS importedKey.algorithm.length is 256
PASS importedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "AES-GCM"
PASS clonedKey.algorithm.length is 256
PASS clonedKey.usages.join(',') is "encrypt,wrapKey,unwrapKey"
Serialized key bytes: 5c4b010920c2012000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0fa0000000014b


PASS successfullyParsed is true

TEST COMPLETE
